refactor: SingleMediaViewer()を Viewer.ktから新ファイルに切り出す
